refactor: move AdminDashboard Component
